Behaviors basics included via module.
Instead of subclassing a base class to get basic behavior functionality (simple_name and notifying service registry of creation), include a module.  This deprecates the original base class for any users currently using inheritance.

Note that the protocol for behaviors is that they must accept constructor arguments as well, a feature which is not captured by either the inheritance or module strategy.